Move the socket.io client singleton to Utils dir.
